move datestring function to backend utils
